add email domain table show/hide and filtering
